browser: Move TraceKit to a separate folder
