+++ packages/browser/src/tracekit/index.js
